Yet again tackling resizing, this time letting flex box do the hard work.
This time, we had issues with the intrinsic aspect ratio from the view box messing up the calculated width and height. Setting the view box to ‘0 0 1 1’ and then measuring the width/height meant that it would try to expand to a 1:1 aspect ratio and then calculate the hard minimum sizes and set them, often leading to plots that would overflow their containing box.

Thanks to @ssunkara1 for working on this commit as well.