Attributes update refactoring
Remove GLAttribute class which makes lots of copies and recreate each and every attribute buffer
on every update (e.g. changing the `x` value was triggering an `update_geometry` which was triggering
the creation of the `x`, `x_previous`, `y`, `y_previous`, `color`, `size`, `size_previous`, `rotation`,
`rotation_previous`... buffers). Creating a new AttributeBuffer means that it needs to be sent to the
GPU, and that is a slow operation.

Now the AttributeBuffers are created only once, and changing the `x`
value only triggers a change on the `x` and `x_previous` AttributeBuffers.