Render the plot on first after-show event
This delays the rendering of the plot to when the plot is actually
visible on the page and has available space.

This fixes an issue when having bqplots in Tab widgets.

Because we waited for the displayed promise to be resolved (which
is resolved on after-attach), the plot was rendered under invisible
tabs with an available size of 0, and was not resized when clicking
on the tab. This was resulting in a weirdly shrinked plot.

This changes this behavior by waiting for the element to be actually
shown on the page (after-show is triggered when clicking on the tab) and
has available space.

Signed-off-by: martinRenou <martin.renou@gmail.com>