feat: enable deduplication of d3-selection for shared browser events
d3-selection has global state (the event) that is obtained
using `require("d3-selection").event`. If multiple copies of this
js module exists in the frontend, they will not read eachother's
events.

This is required for 3rd party libraries, such as bqplot-image-gl
to be able to read the last d3 event.

Since Jupyter Lab will deduplicate direct dependencies, we only need
to add the d3-selection as an explicit dependency, and it will be
shared with other extensions.