fix: axis formatting depended on initial values or changes
The codepath to to axis formatting for the initial values was
different from when properties such as tick_values are updated
afterwards.
This was probably due to some technical debt/circular dependency
in the code.
If did the minimal amount of change to make sure the behaviour is
the same, while having a consistent codepath.