fix: background color of figure was not computed properly (#1603)
On the screen, it is not the computed background color that is seen
by a user. Since DOM elements can be transparent, the background
color might be one of the parent elements. This is why we need to
traverse the DOM tree to find the background color of the first
non-transparent parent element.

A more sophisticated solution would be to also use blending when
there is a non-fully transparent background color. I think that
might be a bit overkill for now.

See also https://github.com/spacetelescope/jdaviz/pull/2264#issuecomment-1612052525